feat: implement Subject CRUD with Tauri managed state
- Fix get_connection to resolve app_data_dir instead of a relative path
- Enable PRAGMA foreign_keys on connection open
- Move connection to managed Mutex<Connection> state in setup()
- Add create_subject, update_subject, delete_subject commands
- Use named column access instead of positional row.get()
- Fix Subject.s_id type from i32 to i64 to match SQLite INTEGER PK
- Guard update/delete against zero rows_affected

+++ src-tauri/src/models/subject.rs
@@ -2,7 +2,7 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct Subject {
-    pub s_id: i32,
+    pub s_id: i64,
     pub code: String,
     pub name: String,
     pub color: String,
